+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+In next release ...
+
+- Features:
+
+  The template loader class now accepts an optional argument
+  ``default_extension`` which accepts a filename extension which will
+  be appended to the filename if there's not already an extension.
+
 2.3.3 (2011-08-15)
 ------------------